refactor: fix plugins folder name
Resolve spelling mistake in the plugins folder name
